modified exception & errorcodes
