+++ src/main/kotlin/com/waffletoy/team1server/exceptions/ApiException.kt
@@ -6,41 +6,6 @@ open class ApiException(
     val details: Map<String, Any>? = null,
 ) : RuntimeException(message)
 
-class UserDuplicateSnuMailException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_DUPLICATE_SNUMAIL,
-    details = details,
-)
-
-class UserDuplicateLocalIdException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_DUPLICATE_LOCAL_ID,
-    details = details,
-)
-
-class UserDuplicateGoogleIdException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_DUPLICATE_GOOGLE_ID,
-    details = details,
-)
-
-class UserNotFoundException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_NOT_FOUND,
-    details = details,
-)
-
-class UserRoleConflictException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_INVALID_ROLE,
-    details = details,
-)
-
-class UserMethodConflictException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_INVALID_METHOD,
-    details = details,
-)
-
-class UserMergeUnknownFailureException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_MERGE_UNKNOWN_FAILURE,
-    details = details,
-)
-
 class BadAuthorizationHeaderException(details: Map<String, Any>? = null) : ApiException(
     errorCode = ErrorCode.AUTH_BAD_HEADER,
     details = details,
@@ -76,47 +41,12 @@ class NotAuthorizedException(details: Map<String, Any>? = null) : ApiException(
     details = details,
 )
 
-class EmailVerificationFailedException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.EMAIL_VERIFICATION_FAILED,
-    details = details,
-)
-
-class EmailVerificationInvalidException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.EMAIL_VERIFICATION_INVALID,
-    details = details,
-)
-
-class EmailVerificationSendFailureException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.EMAIL_VERIFICATION_SEND_FAILURE,
-    details = details,
-)
-
 class NotImplementedException(details: Map<String, Any>? = null) : ApiException(
     errorCode = ErrorCode.NOT_IMPLEMENTED,
     details = details,
 )
 
 class InvalidRequestException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.INVALID_REQUEST,
-    details = details,
-)
-
-class S3UrlGenerationFailedException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.POST_S3_URL_GENERATION_FAILED,
-    details = details,
-)
-
-class S3SDKClientFailedException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.POST_S3_SDK_FAILED,
-    details = details,
-)
-
-class UserNotAuthorizedException(details: Map<String, Any>? = null) : ApiException(
-    errorCode = ErrorCode.USER_INVALID_ROLE,
-    details = details,
-)
-
-class UserSocialLoginInvalidException(details: Map<String, Any>? = null) : ApiException(
     errorCode = ErrorCode.INVALID_REQUEST,
     details = details,
 )
