+++ src/main/kotlin/com/waffletoy/team1server/post/service/S3Service.kt
@@ -4,11 +4,10 @@ import com.amazonaws.HttpMethod
 import com.amazonaws.SdkClientException
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.AmazonS3Exception
-import com.waffletoy.team1server.exceptions.S3SDKClientFailedException
-import com.waffletoy.team1server.exceptions.S3UrlGenerationFailedException
-import com.waffletoy.team1server.exceptions.UserNotAuthorizedException
+import com.waffletoy.team1server.post.*
 import com.waffletoy.team1server.post.controller.PreSignedDownloadReq
 import com.waffletoy.team1server.post.controller.PreSignedUploadReq
+import com.waffletoy.team1server.user.UserInvalidRoleException
 import com.waffletoy.team1server.user.UserRole
 import com.waffletoy.team1server.user.dtos.User
 import org.springframework.beans.factory.annotation.Autowired
@@ -29,16 +28,16 @@ class S3Service(
         expirationMinutes: Long = EXPIRATION_MINUTES,
     ): String {
         if (user.userRole != UserRole.CURATOR) {
-            throw UserNotAuthorizedException()
+            throw UserInvalidRoleException()
         }
         try {
             val filePath = "${preSignedUploadReq.fileName}.${preSignedUploadReq.fileType}"
             val expiration = calculateExpiration(expirationMinutes)
             return amazonS3.generatePresignedUrl(bucketName, filePath, expiration, HttpMethod.PUT).toString()
         } catch (e: AmazonS3Exception) {
-            throw S3UrlGenerationFailedException()
+            throw PostS3UrlGenerationFailedException()
         } catch (e: SdkClientException) {
-            throw S3SDKClientFailedException()
+            throw PostS3SDKClientFailedException()
         }
     }
 
@@ -49,15 +48,15 @@ class S3Service(
         expirationMinutes: Long = EXPIRATION_MINUTES,
     ): String {
         if (user.userRole != UserRole.CURATOR) {
-            throw UserNotAuthorizedException()
+            throw UserInvalidRoleException()
         }
         try {
             val expiration = calculateExpiration(expirationMinutes)
             return amazonS3.generatePresignedUrl(bucketName, preSignedDownloadReq.fileName, expiration, HttpMethod.GET).toString()
         } catch (e: AmazonS3Exception) {
-            throw S3UrlGenerationFailedException()
+            throw PostS3UrlGenerationFailedException()
         } catch (e: SdkClientException) {
-            throw S3SDKClientFailedException()
+            throw PostS3SDKClientFailedException()
         }
     }
